Ensure the "output" directory exists when testing.
There are currently no files in the `lib/src/test/resources/output`
directory, so that directory may not exist (and will not exist) when the
test suite is run. So for now we must ensure the directory exists when
we run the test for the IDRangePolicyWriter, since it needs to write its
output there.