Change command line syntax.
Remove the `--input` option. Instead, we expect the input file to be the
first (and only) positional argument on the command line. When writing
the policy, the default behaviour is to write it back to that same file,
unless the `--output` option is used to specify a different file.

Writing of the policy is automatically triggered if the policy has been
modified (i.e. if `--add-range`) has been used, but it can also be
explicitly requested with the `--save` option.