Add OutOfIDSpaceException.
Add a new exception class to represent any problem resulting from a lack
of available ID space.

Make the IDRangePolicy#addRange() method return such an exception when
it cannot find an open range wide enough for the requested range.

Also create the base IDException class, parent to both the new
OutOfIDSpaceException and the existing InvalidIDRangePolicyException, as
well as a (for now unused) NoSuchIDRangeException.