Throws an exception when the requested range is not found.
Change the behaviour of the IDRangePolicy#getRange() method to throw a
NoSuchIDRangeException when the requested range is not found.

Add a new method IDRangePolicy#findRange() method with the former
behaviour of the #getRange() method -- that is, it returns an
Optional<IDRange> object, and returns an empty Optional if the range is
not found instead of throwing an exception.