Use subcommands.
Lay the groundwork for allowing to use the dicer-cli tool for more tasks
than manipulating a ID range policy file.

The SimpleCLI class is turned into a main command that, in itself, does
nothing, but that can accept several subcommands -- the only subcommand,
for now, being a `policy` command that does what the former `dicer-cli`
command used to do.

The CommandHelper class is no longer necessary as its methods are now
directly part of the SimpleCLI class.