Rename OutOfIDSpaceException.
Rename the OutOfIDSpaceException class to IDNotFoundException. This is
because, depending on how the ID generators (for now the only throwers
of this exception) work, there could be other reason for a failure to
generate an ID than a lack of available ID space.

We now consider that not being able to generate an ID is ackin to not
"finding" an ID. This will also make it possible to maybe reuse that
exception for slightly different error scenarios later.