Allow initializing an ID generator from an ID range.
Add a constructor to both SequentialIDGenerator and
RandomizedIDGenerator that takes an IDRange object.

This is similar to previously removed constructors that accepted
simultaneously either an ID policy and a range name, or a format string
and an IDRange, except that now the IDRange object is self-sufficient as
it provides access to the format string.