Implement the basic functionality of the `tsv` tool.
The `tsv` tool can now inject automatically generated IDs into a TSV
file, which is what the tool is all about.

Still critically missing is the ability to provide the tool with a
source of IDs already in use (typically a OWL ontology), so that it will
only generate _new_ IDs.