Fix 'tsv policy --output...'.
The expected behaviour with `tsv policy` is that specifying an output
file (with `--output <FILE>`) should be enough to trigger writing the
policy back to the specified file, _without_ having to also use the
`--save` option.

This was not working properly because PicoCLI is setting the
`IOOptions#write` field to its default value after we forcibly set it
when processing `--output`.

So we add a new `forceWrite` field, which we can set independently of
the PicoCLI-managed `write` field. We enquire about whether we should
write the policy or not with a new method that return `forceWrite` if it
is true (overriding the PicoCLI-managed `write`), or `write` otherwise.