Make the policy command more useful in case of errors.
When a policy file is syntactically incorrect (i.e. it cannot be loaded
by the OWLAPI), `dicer-cli policy` only shows a generic "Cannot load ID
policy file" message, which is of no use to help users finding and
fixing the problem.

So we add two options to make the output more useful:

* `--show-owlapi-error`, to make the command show the full error message
  from the OWLAPI;
* `--assume-manchester`, to try loading the file using _only_ the
  Manchester parser (based on the fact that most if not all OBO ID
  policy files are in Manchester syntax -- at least all the
  ODK-generated ones) -- this will produce an error message containing
  only the real error found by the Manchester parser.